feat: Order Entity에 비즈니스 로직 추가

+++ src/main/java/org/prgms/kdt/OrderItem.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 public class OrderItem {
     public final UUID productId;
     public final long productPrice;
-    public final int quantity;
+    public final long quantity;
 
     public OrderItem(UUID productId, long productPrice, int quantity) {
         this.productId = productId;
@@ -21,7 +21,7 @@ public class OrderItem {
         return productPrice;
     }
 
-    public int getQuantity() {
+    public long getQuantity() {
         return quantity;
     }
 }
